% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_data.R
\docType{data}
\name{colors_default}
\alias{colors_default}
\title{Standaard kleurenpalet voor NFWA visualisaties}
\format{
Een named character vector met hex-codes en kleurnamen:
\describe{
  \item{title_color}{Kleur voor hoofdtitels (zwart)}
  \item{positive_color}{Kleur voor positieve waarden (blauw)}
  \item{negative_color}{Kleur voor negatieve waarden (rood)}
  \item{metrics_green}{Kleur voor groene metrieken}
  \item{metrics_red}{Kleur voor rode metrieken}
  \item{color_bias_positive}{Kleur voor positieve bias (groen)}
  \item{color_bias_negative}{Kleur voor negatieve bias (roze)}
  \item{color_bias_neutral}{Kleur voor neutrale bias (oranje)}
  \item{color_bias_none}{Kleur voor geen bias (grijs)}
}
}
\usage{
colors_default
}
\description{
Een named vector met standaardkleuren voor titels, tekst, achtergronden,
gridlijnen, en fairness-indicatoren.
}
\keyword{internal}
