% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_data.R
\docType{data}
\name{colors_list}
\alias{colors_list}
\title{Specifieke kleurenpaletten voor sensitieve variabelen}
\format{
Een named list met kleurvectoren:
\describe{
  \item{geslacht}{Kleuren voor geslacht (M, V)}
  \item{vooropleiding}{Kleuren voor vooropleiding (MBO, HAVO, VWO, etc.)}
  \item{aansluiting}{Kleuren voor type aansluiting (Direct, Tussenjaar, etc.)}
  \item{roc_plots}{Kleuren voor ROC-curve plots}
}
}
\usage{
colors_list
}
\description{
Een named list met kleurvectoren voor verschillende categorische variabelen
zoals geslacht, vooropleiding en aansluiting.
}
\keyword{datasets}
