% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_explain_lf.R
\name{create_explain_lf}
\alias{create_explain_lf}
\title{Maak een DALEX-explainer van een afgerond model}
\usage{
create_explain_lf(df, last_fit, best_model)
}
\arguments{
\item{df}{Verwerkte 1CHO en 1CHO vak data uit NFWA package}

\item{last_fit}{Een `last_fit`-object uit het tidymodels-framework, met
daarin het getrainde model en de bijbehorende workflow.}

\item{best_model}{Character. Naam van het beste model, gebruikt als label
voor de explainer (bijv. `"Logistic Regression"`).}
}
\value{
Een [DALEX::explain()]-object dat kan worden gebruikt voor
  fairness-checks en model-interpretatie.
}
\description{
Extraheert het getrainde model en de workflow uit een `last_fit`-object
en maakt een DALEX-explainer aan waarmee fairness-analyses kunnen worden
uitgevoerd.
}
\keyword{internal}
