% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_largest_group.R
\name{get_largest_group}
\alias{get_largest_group}
\title{Bepaal de grootste (geprivilegieerde) groep}
\usage{
get_largest_group(df, var)
}
\arguments{
\item{df}{Data frame met de data.}

\item{var}{Character. Naam van de variabele waarvan de grootste groep
bepaald moet worden.}
}
\value{
Character. Naam van de meest voorkomende subgroep.
}
\description{
Identificeert de meest voorkomende subgroep binnen een variabele.
Deze groep wordt gebruikt als referentie (geprivilegieerde groep) in
fairness-analyses. Bij gelijke aantallen wordt de eerste gekozen.
}
